Well, it's a one line fix. So don't be mislead by the length of the ChangeLog entry ;-)
003-03-06  Arvind Samptur <arvind.samptur@wipro.com>
        * panel-properties-dialog.c :
        (panel_properties_dialog_update_background_opacity) : When
        recalculating the percentage, cast it to double. Else you always
        land up with one value less because of integer division and
        that spins off the continuous scale setting. Fixes #107608
        Thanks to Satyajit for fighting this one.

+++ gnome-panel/panel-properties-dialog.c
@@ -579,7 +579,7 @@ panel_properties_dialog_update_background_opacity (PanelPropertiesDialog *dialog
 	if (!value || value->type != GCONF_VALUE_INT)
 		return;
 
-	percentage = (gconf_value_get_int (value) * 100) / 65535;
+	percentage = ((gdouble) (gconf_value_get_int (value) * 100)) / 65535;
 
 	if ((int) gtk_range_get_value (GTK_RANGE (dialog->opacity_scale)) != (int) percentage)
 		gtk_range_set_value (GTK_RANGE (dialog->opacity_scale), percentage);
